Reset store, move compareEvents and add test
